Fix default state inconsistent with README
The README states that, in absence of an `imageToolsEnabled` setting in
the options, the plugin is **disabled** by default. It states the same
for `imageToolsOpen`.

However, the actual behavior was the opposite of that: Due to the
presence of a `true` defaultProp for both values, they would be
**enabled** in the absence of their respective config key.

This commit fixes this inconsistency by explicitely setting `enabled`
and `open` to `false` if the configuration key is absent (i.e.
`undefined`).